+++ internal/pkg/util/bin/bin.go
@@ -14,9 +14,9 @@ import (
 	"fmt"
 	"os"
 	"os/exec"
+	"strings"
 
 	"github.com/apptainer/apptainer/internal/pkg/buildcfg"
-	"github.com/apptainer/apptainer/internal/pkg/util/env"
 	"github.com/apptainer/apptainer/pkg/sylog"
 	"github.com/apptainer/apptainer/pkg/util/apptainerconf"
 	"github.com/pkg/errors"
@@ -33,28 +33,42 @@ func FindBin(name string) (path string, err error) {
 		return findOnPath(name)
 	// Configurable executables that are found at build time, can be overridden
 	// in apptainer.conf. If config value is "" will look on PATH.
-	case "unsquashfs", "mksquashfs", "go":
+	case "unsquashfs", "mksquashfs", "go", "cryptsetup", "ldconfig", "nvidia-container-cli":
 		return findFromConfigOrPath(name)
 	// distro provided setUID executables that are used in the fakeroot flow to setup subuid/subgid mappings
 	case "newuidmap", "newgidmap":
 		return findOnPath(name)
-	// cryptsetup & nvidia-container-cli paths must be explicitly specified
-	// They are called as root from the RPC server in a setuid install, so this
-	// limits to sysadmin controlled paths.
-	// ldconfig is invoked by nvidia-container-cli, so must be trusted also.
-	case "cryptsetup", "ldconfig", "nvidia-container-cli":
-		return findFromConfigOnly(name)
 	}
 	return "", fmt.Errorf("unknown executable name %q", name)
 }
 
-// findOnPath performs a simple search on PATH for the named executable, returning its full path.
-// env.DefaultPath` is appended to PATH to ensure standard locations are searched. This
-// is necessary as some distributions don't include sbin on user PATH etc.
+// findOnPath performs a search on the configurated binary path for the
+// named executable, returning its full path.
 func findOnPath(name string) (path string, err error) {
+	cfg := apptainerconf.GetCurrentConfig()
+	if cfg == nil {
+		if strings.HasSuffix(os.Args[0], ".test") {
+			// read config if doing unit tests
+			cfg, err = apptainerconf.Parse(buildcfg.APPTAINER_CONF_FILE)
+			if err != nil {
+				return "", errors.Wrap(err, "unable to parse apptainer configuration file")
+			}
+			apptainerconf.SetCurrentConfig(cfg)
+		} else {
+			sylog.Fatalf("configuration not pre-loaded in findOnPath")
+		}
+	}
+	newPath := cfg.BinaryPath
+	if strings.Contains(newPath, "$PATH:") {
+		if strings.HasSuffix(os.Args[0], ".test") {
+			apptainerconf.SetBinaryPath("", true)
+		} else {
+			sylog.Fatalf("SetBinaryPath has not been run before findOnPath")
+		}
+	}
 	oldPath := os.Getenv("PATH")
 	defer os.Setenv("PATH", oldPath)
-	os.Setenv("PATH", oldPath+":"+env.DefaultPath)
+	os.Setenv("PATH", newPath)
 
 	path, err = exec.LookPath(name)
 	if err != nil {
@@ -68,9 +82,15 @@ func findOnPath(name string) (path string, err error) {
 func findFromConfigOrPath(name string) (path string, err error) {
 	cfg := apptainerconf.GetCurrentConfig()
 	if cfg == nil {
-		cfg, err = apptainerconf.Parse(buildcfg.APPTAINER_CONF_FILE)
-		if err != nil {
-			return "", errors.Wrap(err, "unable to parse apptainer configuration file")
+		if strings.HasSuffix(os.Args[0], ".test") {
+			// read config if doing unit tests
+			cfg, err = apptainerconf.Parse(buildcfg.APPTAINER_CONF_FILE)
+			if err != nil {
+				return "", errors.Wrap(err, "unable to parse apptainer configuration file")
+			}
+			apptainerconf.SetCurrentConfig(cfg)
+		} else {
+			sylog.Fatalf("configuration not pre-loaded in findFromConfigOrPath")
 		}
 	}
 
@@ -81,32 +101,6 @@ func findFromConfigOrPath(name string) (path string, err error) {
 		path = cfg.MksquashfsPath
 	case "unsquashfs":
 		path = cfg.UnsquashfsPath
-	default:
-		return "", fmt.Errorf("unknown executable name %q", name)
-	}
-
-	if path == "" {
-		return findOnPath(name)
-	}
-
-	sylog.Debugf("Using %q at %q (from apptainer.conf)", name, path)
-
-	// Use lookPath with the absolute path to confirm it is accessible & executable
-	return exec.LookPath(path)
-}
-
-// findFromConfigOnly retrieves the path to an executable from apptainer.conf.
-// If it's not set there we error.
-func findFromConfigOnly(name string) (path string, err error) {
-	cfg := apptainerconf.GetCurrentConfig()
-	if cfg == nil {
-		cfg, err = apptainerconf.Parse(buildcfg.APPTAINER_CONF_FILE)
-		if err != nil {
-			return "", errors.Wrap(err, "unable to parse apptainer configuration file")
-		}
-	}
-
-	switch name {
 	case "cryptsetup":
 		path = cfg.CryptsetupPath
 	case "ldconfig":
@@ -118,7 +112,7 @@ func findFromConfigOnly(name string) (path string, err error) {
 	}
 
 	if path == "" {
-		return "", fmt.Errorf("path to %q not set in apptainer.conf", name)
+		return findOnPath(name)
 	}
 
 	sylog.Debugf("Using %q at %q (from apptainer.conf)", name, path)